requirements.txt
asgiref==3.3.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==3.3
inflection==0.5.1
instagrapi==1.15.11
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow>=8.1.1
psycopg2-binary==2.9
pycparser==2.20
pycryptodomex==3.9.9
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.3
PySocks==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
social-auth-app-django==5.0.0
social-auth-core @ git+https://github.com/senabo/social-core.git@9acf09a988b705f1ba32c56d0c05f65be16488f2
sqlparse==0.3.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
